ExecutionService の投影更新を ExecutionProjectionOrchestrator に分離し、queue と operational sync の窓口を一本化する。
