+++ core/application/Statevia.Core.Application/DependencyInjection/ApplicationServiceCollectionExtensions.cs
@@ -26,6 +26,7 @@ public static class ApplicationServiceCollectionExtensions
         services.AddSingleton<ExecutionOwnershipTracker>();
         services.AddScoped<ExecutionQueryService>();
         services.AddScoped<ExecutionIdempotencyService>();
+        services.AddScoped<ExecutionProjectionOrchestrator>();
         services.AddScoped<IExecutionService, ExecutionService>();
         services.AddScoped<IEventIngressService, EventIngressService>();
         services.AddScoped<IForkChildExecutionCoordinator, ForkChildExecutionCoordinator>();
